@@ -12,17 +12,21 @@
     "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
-    "fullcalendar-scheduler": "^1.5.0",
+    "fullcalendar": "file:src/lib/fullcalendar",
+    "fullcalendar-scheduler": "1.5.0",
     "jquery": "^3.1.1",
     "jquery.nicescroll": "^3.6.8",
     "moment": "^2.17.1",
-    "react": "^15.4.1",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
     "react-addons-css-transition-group": "^15.4.1",
+    "react-addons-shallow-compare": "^15.5.2",
     "react-addons-update": "^15.4.1",
     "react-dom": "^15.4.1",
-    "react-infinite-calendar": "^1.2.2",
+    "react-infinite-calendar": "^2.1.6",
     "react-redux": "^4.4.6",
-    "react-router": "^4.0.0-alpha.6",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
     "react-select": "^1.0.0-rc.2",
     "react-slick": "^0.14.5",
     "redux": "^3.6.0",
