@@ -1,22 +1,29 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import Select from 'react-select';
 import * as Images from '../../../../require/images';
 
 /* Search */
 
-const OptionComponent = React.createClass({
+class OptionComponent extends React.Component {
+	constructor (props) {
+		super (props);
+		this.handleMouseDown = this.handleMouseDown.bind(this);
+		this.handleMouseEnter = this.handleMouseEnter.bind(this);
+		this.handleMouseMove = this.handleMouseMove.bind(this);
+	}
 	handleMouseDown (event) {
 		event.preventDefault();
 		event.stopPropagation();
 		this.props.onSelect(this.props.option, event);
-	},
+	}
 	handleMouseEnter (event) {
 		this.props.onFocus(this.props.option, event);
-	},
+	}
 	handleMouseMove (event) {
 		if (this.props.isFocused) return;
 		this.props.onFocus(this.props.option, event);
-	},
+	}
 	render () {
 		return (
 			<div className={this.props.className}
@@ -34,9 +41,9 @@ const OptionComponent = React.createClass({
 			</div>
 		);
 	}
-});
+}
 
-const ValueComponent = React.createClass({
+class ValueComponent extends React.Component {
 	render () {
 		return (
 			<div className="Select-value">
@@ -48,25 +55,26 @@ const ValueComponent = React.createClass({
 			</div>
 		);
 	}
-});
+}
 
-const SearchCustomer = React.createClass({
-	propTypes: {
-		hint: React.PropTypes.string,
-		label: React.PropTypes.string,
-	},
-	getInitialState () {
-		return {};
-	},
+class SearchCustomer extends React.Component {
+	constructor (props) {
+		super(props);
+		this.state = {
+		}
+		this.setValue = this.setValue.bind(this);
+	}
+
 	setValue (value) {
 		this.setState({ value });
     this.props.onChange(value);
-	},
+	}
+
   arrowRenderer () {
   	return (
   		<span>+</span>
   	);
-  },
+  }
 
 	render () {
 		return (
@@ -86,7 +94,11 @@ const SearchCustomer = React.createClass({
 			</div>
 		);
 	}
-});
+}
 
+SearchCustomer.propTypes = {
+	hint: PropTypes.string,
+	label: PropTypes.string,
+}
 
 module.exports = SearchCustomer;
