@@ -1,22 +1,29 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import Select from 'react-select';
 import * as Functions from '../../../../js/common';
 import * as Images from '../../../../require/images';
 
 /* Search */
-const OptionComponent = React.createClass({
+class OptionComponent extends React.Component {
+	constructor (props) {
+		super(props);
+		this.handleMouseDown = this.handleMouseDown.bind(this);
+		this.handleMouseEnter = this.handleMouseEnter.bind(this);
+		this.handleMouseMove = this.handleMouseMove.bind(this);
+	}
 	handleMouseDown (event) {
 		event.preventDefault();
 		event.stopPropagation();
 		this.props.onSelect(this.props.option, event);
-	},
+	}
 	handleMouseEnter (event) {
 		this.props.onFocus(this.props.option, event);
-	},
+	}
 	handleMouseMove (event) {
 		if (this.props.isFocused) return;
 		this.props.onFocus(this.props.option, event);
-	},
+	}
 	render () {
 		return (
 			<div
@@ -34,9 +41,9 @@ const OptionComponent = React.createClass({
 			</div>
 		);
 	}
-});
+}
 
-const ValueComponent = React.createClass({
+class ValueComponent extends React.Component {
 	render () {
 		return (
 			<div className="Select-value">
@@ -48,12 +55,9 @@ const ValueComponent = React.createClass({
 			</div>
 		);
 	}
-});
+}
 
-const GenderFilterComponent = React.createClass({
-	propTypes: {
-		onChange: React.PropTypes.func.isRequired
-	},
+class GenderFilterComponent extends React.Component {
 	render () {
 		return (
 			<div className="Select-filter-wrap">
@@ -95,15 +99,18 @@ const GenderFilterComponent = React.createClass({
 			</div>
 		);
 	}
-})
+}
 
-
-
-const SearchProduct = React.createClass({
-	propTypes: {
-		hint: React.PropTypes.string,
-		label: React.PropTypes.string,
-	},
+class SearchProduct extends React.Component {
+	constructor (props) {
+		super (props);
+		this.state = {
+			genderCode: 0,
+			value: this.props.value
+		}
+		this.setGenderCode = this.setGenderCode.bind(this);
+		this.setValue = this.setValue.bind(this);
+	}
 	componentDidMount () {
 		let _component = this;
 		if (this.props.autoDropdown) {
@@ -111,7 +118,7 @@ const SearchProduct = React.createClass({
 				_component.dropDownSelectOptions();
 			},0);
 		}
-	},
+	}
 	// react-select 모듈의 옵션에 초기 자동 드롭다운 옵션이 없으므로 트리거하여 드롭다운 하도록 함수
 	dropDownSelectOptions () {
 		this.refs.select.handleMouseDown({
@@ -119,27 +126,22 @@ const SearchProduct = React.createClass({
 			preventDefault: function () {},
 			stopPropagation: function () {}
 		});
-	},
-	getInitialState () {
-		return {
-			genderCode: 0,
-			value: this.props.value
-		};
-	},
+	}
+
 	setGenderCode (value) {
 		this.setState({
 			genderCode: value
 		});
-	},
+	}
 	setValue (value) {
 		this.setState({ value });
     this.props.onChange(value);
-	},
+	}
   arrowRenderer () {
   	return (
   		<span>+</span>
   	);
-  },
+  }
 	getGenderCode(productObj, inputStr){
 		let code = productObj.sex *1;
 
@@ -158,7 +160,7 @@ const SearchProduct = React.createClass({
 			default :
 				return false;
 		}
-	},
+	}
 
 	render () {
 		{/*filterOption={this.getGenderCode}*/}
@@ -189,7 +191,16 @@ const SearchProduct = React.createClass({
 			</div>
 		);
 	}
-});
+}
+
+GenderFilterComponent.propTypes = {
+	onChange: PropTypes.func.isRequired
+}
+
+SearchProduct.propTypes = {
+	hint: PropTypes.string,
+	label: PropTypes.string,
+}
 
 
 module.exports = SearchProduct;
